+++ driver-core/src/main/java/com/datastax/driver/core/Row.java
@@ -405,6 +405,50 @@ public interface Row {
     public InetAddress getInet(String name);
 
     /**
+     * Returns the {@code i}th value of this row as a {@link Token}.
+     *
+     * @param i the index ({@code 0 <= i < size()}) of the column to retrieve.
+     * @return the value of the {@code i}th column in this row as an Token.
+     * If the value is NULL, {@code null} is returned.
+     *
+     * @throws IndexOutOfBoundsException if {@code i < 0 || i >= this.columns().size()}.
+     * @throws InvalidTypeException if column {@code i} is not of the type of token values
+     * for this cluster (this depends on the configured partitioner).
+     */
+    public Token getToken(int i);
+
+    /**
+     * Returns the value of column {@code name} as a {@link Token}.
+     * <p>
+     * This method will first try with {@code token(name)}, because in most cases the column
+     * is named that way since it results from the application of a CQL function:
+     * <pre>
+     * {@code
+     * ResultSet rs = session.execute("SELECT token(k) FROM my_table WHERE k = 1");
+     * Token token = rs.one().getToken("k"); // retrieves token(k)
+     * }
+     * </pre>
+     * If that fails, it will try will {@code name}. This handles the case where the column
+     * was aliased:
+     * <pre>
+     * {@code
+     * ResultSet rs = session.execute("SELECT token(k) AS t FROM my_table WHERE k = 1");
+     * Token token = rs.one().getToken("t");
+     * }
+     * </pre>
+     *
+     * @param name the name of the column to retrieve.
+     * @return the value of column {@code name} as a Token.
+     * If the value is NULL, {@code null} is returned.
+     *
+     * @throws IllegalArgumentException if {@code name} is not part of the
+     * ResultSet this row is part of, i.e. if {@code !this.columns().names().contains(name)}.
+     * @throws InvalidTypeException if column {@code name} is not of the type of token values
+     * for this cluster (this depends on the configured partitioner).
+     */
+    public Token getToken(String name);
+
+    /**
      * Returns the {@code i}th value of this row as a list.
      *
      * @param <T> the type of the elements of the list to return.