+++ driver-core/src/test/java/com/datastax/driver/core/OPPTokenVnodeIntegrationTest.java
@@ -1,0 +1,11 @@
+package com.datastax.driver.core;
+
+public class OPPTokenVnodeIntegrationTest extends TokenIntegrationTest {
+    public OPPTokenVnodeIntegrationTest() {
+        super("-p ByteOrderedPartitioner", DataType.blob(), true);
+    }
+
+    @Override protected Token.Factory tokenFactory() {
+        return Token.OPPToken.FACTORY;
+    }
+}